Make the AST interpret itself
Replace the hash AST with one that consists of node objects. Each AST node can evaluate itself, which means that the job of TreeInterpreter is replaced by calling #visit on the top node in the AST.
