+++ lib/jmespath.rb
@@ -7,11 +7,11 @@ module JMESPath
   autoload :Errors, 'jmespath/errors'
   autoload :ExprNode, 'jmespath/expr_node'
   autoload :Lexer, 'jmespath/lexer'
+  autoload :Nodes, 'jmespath/nodes'
   autoload :Parser, 'jmespath/parser'
   autoload :Runtime, 'jmespath/runtime'
   autoload :Token, 'jmespath/token'
   autoload :TokenStream, 'jmespath/token_stream'
-  autoload :TreeInterpreter, 'jmespath/tree_interpreter'
   autoload :VERSION, 'jmespath/version'
 
   class << self
