+++ lib/jmespath/nodes/index.rb
@@ -1,0 +1,27 @@
+module JMESPath
+  # @api private
+  module Nodes
+    class Index < Node
+      attr_reader :index
+
+      def initialize(index)
+        @index = index
+      end
+
+      def visit(value)
+        if Array === value
+          value[@index]
+        else
+          nil
+        end
+      end
+
+      def to_h
+        {
+          :type => :index,
+          :index => @index,
+        }
+      end
+    end
+  end
+end
