+++ lib/jmespath/parser.rb
@@ -15,7 +15,7 @@ module JMESPath
         :filter,            # foo.[?bar==10]
     ])
 
-    CURRENT_NODE = { type: :current }
+    CURRENT_NODE = Nodes::Current.new
 
     # @option options [Lexer] :lexer
     def initialize(options = {})
@@ -61,10 +61,7 @@ module JMESPath
 
     def nud_expref(stream)
       stream.next
-      {
-        type: :expression,
-        children: [expr(stream, 2)]
-      }
+      Nodes::Expression.new([expr(stream, 2)])
     end
 
     def nud_filter(stream)
@@ -78,7 +75,7 @@ module JMESPath
     def nud_identifier(stream)
       token = stream.token
       stream.next
-      { type: :field, key: token.value }
+      Nodes::Field.new(token.value)
     end
 
     def nud_lbrace(stream)
@@ -92,10 +89,7 @@ module JMESPath
         end
       end while stream.token.type != :rbrace
       stream.next
-      {
-        type: :multi_select_hash,
-        children: pairs
-      }
+      Nodes::MultiSelectHash.new(pairs)
     end
 
     def nud_lbracket(stream)
@@ -113,10 +107,7 @@ module JMESPath
     def nud_literal(stream)
       value = stream.token.value
       stream.next
-      {
-        type: :literal,
-        value: value
-      }
+      Nodes::Literal.new(value)
     end
 
     def nud_quoted_identifier(stream)
@@ -126,7 +117,7 @@ module JMESPath
         msg = 'quoted identifiers are not allowed for function names'
         raise Errors::SyntaxError, msg
       else
-        { type: :field, key: token[:value] }
+        Nodes::Field.new(token[:value])
       end
     end
 
@@ -137,14 +128,8 @@ module JMESPath
     def led_comparator(stream, left)
       token = stream.token
       stream.next
-      {
-        type: :comparator,
-        relation: token.value,
-        children: [
-          left,
-          expr(stream),
-        ]
-      }
+      children = [left, expr(stream)]
+      Nodes::Comparator.new(children, token.value)
     end
 
     def led_dot(stream, left)
@@ -152,13 +137,8 @@ module JMESPath
       if stream.token.type == :star
         parse_wildcard_object(stream, left)
       else
-        {
-          type: :subexpression,
-          children: [
-            left,
-            parse_dot(stream, Token::BINDING_POWER[:dot])
-          ]
-        }
+        children = [left, parse_dot(stream, Token::BINDING_POWER[:dot])]
+        Nodes::Subexpression.new(children)
       end
     end
 
@@ -170,42 +150,31 @@ module JMESPath
       end
       stream.next
       rhs = parse_projection(stream, Token::BINDING_POWER[:filter])
-      {
-        type: :projection,
-        from: :array,
-        children: [
-          left ? left : CURRENT_NODE,
-          {
-            type: :condition,
-            children: [expression, rhs],
-          }
-        ]
-      }
+      children = [
+        left ? left : CURRENT_NODE,
+        Nodes::Condition.new([expression, rhs])
+      ]
+      Nodes::Projection.new(children, :array)
     end
 
     def led_flatten(stream, left)
       stream.next
-      {
-        type: :projection,
-        from: :array,
-        children: [
-          { type: :flatten, children: [left] },
-          parse_projection(stream, Token::BINDING_POWER[:flatten])
-        ]
-      }
+      children = [
+        Nodes::Flatten.new([left]),
+        parse_projection(stream, Token::BINDING_POWER[:flatten])
+      ]
+      Nodes::Projection.new(children, :array)
     end
 
     def led_lbracket(stream, left)
       stream.next(match: Set.new([:number, :colon, :star]))
       type = stream.token.type
       if type == :number || type == :colon
-        {
-          type: :subexpression,
-          children: [
-            left,
-            parse_array_index_expression(stream)
-          ]
-        }
+        children = [
+          left,
+          parse_array_index_expression(stream)
+        ]
+        Nodes::Subexpression.new(children)
       else
         parse_wildcard_array(stream, left)
       end
@@ -213,7 +182,7 @@ module JMESPath
 
     def led_lparen(stream, left)
       args = []
-      name = left[:key]
+      name = left.key
       stream.next
       while stream.token.type != :rparen
         args << expr(stream, 0)
@@ -222,27 +191,19 @@ module JMESPath
         end
       end
       stream.next
-      {
-        type: :function,
-        fn: name,
-        children: args,
-      }
+      Nodes::Function.new(args, name)
     end
 
     def led_or(stream, left)
       stream.next
-      {
-        type: :or,
-        children: [left, expr(stream, Token::BINDING_POWER[:or])]
-      }
+      children = [left, expr(stream, Token::BINDING_POWER[:or])]
+      Nodes::Or.new(children)
     end
 
     def led_pipe(stream, left)
       stream.next
-      {
-        type: :pipe,
-        children: [left, expr(stream, Token::BINDING_POWER[:pipe])],
-      }
+      children = [left, expr(stream, Token::BINDING_POWER[:pipe])]
+      Nodes::Pipe.new(children)
     end
 
     def parse_array_index_expression(stream)
@@ -258,11 +219,11 @@ module JMESPath
       end while stream.token.type != :rbracket
       stream.next
       if pos == 0
-        { type: :index, index: parts[0] }
+        Nodes::Index.new(parts[0])
       elsif pos > 2
         raise Errors::SyntaxError, 'invalid array slice syntax: too many colons'
       else
-        { type: :slice, args: parts }
+        Nodes::Slice.new(parts)
       end
     end
 
@@ -279,11 +240,7 @@ module JMESPath
       key = stream.token.value
       stream.next(match:Set.new([:colon]))
       stream.next
-      {
-        type: :key_value_pair,
-        key: key,
-        children: [expr(stream)]
-      }
+      Nodes::MultiSelectHash::KeyValuePair.new([expr(stream)], key)
     end
 
     def parse_multi_select_list(stream)
@@ -298,10 +255,7 @@ module JMESPath
         end
       end while stream.token.type != :rbracket
       stream.next
-      {
-        type: :multi_select_list,
-        children: nodes
-      }
+      Nodes::MultiSelectList.new(nodes)
     end
 
     def parse_projection(stream, binding_power)
@@ -321,26 +275,20 @@ module JMESPath
     def parse_wildcard_array(stream, left = nil)
       stream.next(match:Set.new([:rbracket]))
       stream.next
-      {
-        type: :projection,
-        from: :array,
-        children: [
-          left ? left : CURRENT_NODE,
-          parse_projection(stream, Token::BINDING_POWER[:star])
-        ]
-      }
+      children = [
+        left ? left : CURRENT_NODE,
+        parse_projection(stream, Token::BINDING_POWER[:star])
+      ]
+      Nodes::Projection.new(children, :array)
     end
 
     def parse_wildcard_object(stream, left = nil)
       stream.next
-      {
-        type: :projection,
-        from: :object,
-        children: [
-          left ? left : CURRENT_NODE,
-          parse_projection(stream, Token::BINDING_POWER[:star])
-        ]
-      }
+      children = [
+        left ? left : CURRENT_NODE,
+        parse_projection(stream, Token::BINDING_POWER[:star])
+      ]
+      Nodes::Projection.new(children, :object)
     end
 
   end
